==> Scripts/5.merge.runs.EA.Rmd <==
---
title: "merge_runs.Rmd"
author: "Eily Allan"
date: "`r format(Sys.time(), '%B %d, %Y')`"
---

Last updated: 2/2/22

# Overview

We ran dada2 separately for each marker - and we have samples for each month split across multiple sequencing runs. So what we want to do is take the output from dada2 for each run and smash them together so we have one big ASV table for all the samples for a given marker. This means we are going to run this script each time we add another sequencing run. 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(here)
# Check that here() starts at NextGenNEPA
```

# Load datasets and metadata

The ASV tables and hash keys come from the output folder where each run has a subfolder for dada2 and then subfolders for each marker. The metadata actually is stored in the Input folder to keep all the columns that we want (not use dada2 generated metadata file). 


```{r just file paths so far}

# use the pattern to sort out the files - ^ means "starts with" 
all.asvs <- list.files(path = here("Output","dada2_output"), pattern = "^ASV_table.csv", recursive = T, full.names = T)
all.hashes <- list.files(path = here("Output","dada2_output"), pattern = "^Hash_key.csv", recursive = T, full.names = T,ignore.case = T)

all.metadata <- read.csv(here("Input", "sequencing_metadata_files", "master_sequencing_datasheet.csv"))

# now sort ASVs and hashes by marker
COI.asvs <- str_subset(all.asvs, "COI")
COI.hashes <- str_subset(all.hashes, "COI")

MiFish.asvs <- str_subset(all.asvs, "MiFish")
MiFish.hashes <- str_subset(all.hashes, "MiFish")

MiMammal.asvs <- str_subset(all.asvs, "MiMammal")
MiMammal.hashes <- str_subset(all.hashes, "MiMammal")

```

# Merge ASV and hash keys for each marker. 

Now let's actually start pulling ASV and hash files together and merging them all - we will do the metadata after. For some of them, the metadata file names might be wrong and so we need to make sure all the columns and conventions are the same across all the runs we are trying to merge together. 

## Merge COI runs

```{r pull and merge ASV tables and hashes COI}

# coi run 1 has sample as NGNX rather than sample name 
COI1 <- read.csv(COI.asvs[2])
COI1$Sample_name <- as.character(COI1$Sample_name)
COI1 <- COI1 %>% select(-Locus) %>% left_join(all.metadata, by = "Sample_name") %>% select(Sample_ID,Locus,Hash,nReads) %>% rename("Sample_name" = "Sample_ID")

COI2 <- read.csv(COI.asvs[5])

COI3 <- read.csv(COI.asvs[6])

COI4 <- read.csv(COI.asvs[7])
COI4$Sample_name <- as.character(COI4$Sample_name)
COI4 <- COI4 %>% select(-Locus) %>% left_join(all.metadata, by = "Sample_name") %>% select(Sample_ID,Locus,Hash,nReads) %>% rename("Sample_name" = "Sample_ID")

COI5 <- read.csv(COI.asvs[8])

COI6 <- read.csv(COI.asvs[9])
COI6$Sample_name <- as.character(COI6$Sample_name)
COI6 <- COI6 %>% select(-Locus) %>% left_join(all.metadata, by = "Sample_name") %>% select(Sample_ID,Locus,Hash,nReads) %>% rename("Sample_name" = "Sample_ID")

COI7 <- read.csv(COI.asvs[10])
COI7 <- COI7 %>% 
  separate(Sample_name, into=c("marker","date","creek","site","bio_rep"), remove = TRUE) %>%  # separate sample names 
  unite("Sample_name", c("marker","date","creek","site","bio_rep"), sep=".", remove=TRUE)

COI8 <- read.csv(COI.asvs[11])
COI8 <- COI8 %>% 
  separate(Sample_name, into=c("marker","date","creek","site","bio_rep"), remove = TRUE) %>%  # separate sample names 
  unite("Sample_name", c("marker","date","creek","site","bio_rep"), sep=".", remove=TRUE)

COI9 <- read.csv(COI.asvs[12])
COI9 <- COI9 %>% 
  separate(Sample_name, into=c("marker","date","creek","site","bio_rep"), remove = TRUE) %>%  # separate sample names 
  unite("Sample_name", c("marker","date","creek","site","bio_rep"), sep=".", remove=TRUE)

COI10 <- read.csv(COI.asvs[3])
COI10 <- COI10 %>% 
  separate(Sample_name, into=c("marker","date","creek","site","bio_rep"), remove = TRUE) %>%  # separate sample names 
  unite("Sample_name", c("marker","date","creek","site","bio_rep"), sep=".", remove=TRUE)

COI11 <- read.csv(COI.asvs[4])
COI11 <- COI11 %>% 
  separate(Sample_name, into=c("marker","date","creek","site","bio_rep"), remove = TRUE) %>%  # separate sample names 
  unite("Sample_name", c("marker","date","creek","site","bio_rep"), sep=".", remove=TRUE)


COI.ASV.table <- bind_rows(COI1, COI2, COI3, COI4, COI5, COI6, COI7, COI8, COI9, COI10, COI11)

COI.hash.key <- bind_rows(map(COI.hashes, read_csv))
COI.hash.key <- COI.hash.key %>% 
  distinct(Hash, .keep_all = T) 

```

## Merge MiFish runs

```{r pull and merge ASV tables and hashes MiFish}

# we don't want rosetta stone run so skip 1

# we don't want run 1 because we resequenced everything at 35 cycles

MiFish2 <- read.csv(MiFish.asvs[7])
MiFish2 <- MiFish2 %>% 
  separate(Sample_name, into=c("marker","date","creek","site","bio_rep","tech_rep"), remove = TRUE) %>%  # separate sample names 
  mutate(tech_rep = replace_na(tech_rep, 1)) %>% 
    mutate(., tech_rep = case_when(tech_rep == "TR2" ~ "2",
                                 tech_rep == "TR3" ~ "3",
                             TRUE ~ tech_rep)) %>% 
  unite("Sample_name", c("marker","date","creek","site","bio_rep","tech_rep"), sep=".", remove=TRUE) %>% 
  mutate(., Sample_name = case_when(Sample_name == "MiFish.Kangaroo.Run.2.NA.1" ~ "MiFish.Kangaroo.Run2",
                             TRUE ~ Sample_name))

MiFish3 <- read.csv(MiFish.asvs[8])
MiFish3 <- MiFish3 %>% 
  separate(Sample_name, into=c("marker","date","creek","site","bio_rep","tech_rep"), remove = TRUE) %>%  # separate sample names 
  mutate(tech_rep = replace_na(tech_rep, 1)) %>% 
    mutate(., tech_rep = case_when(tech_rep == "TR2" ~ "2",
                                 tech_rep == "TR3" ~ "3",
                             TRUE ~ tech_rep)) %>% 
  unite("Sample_name", c("marker","date","creek","site","bio_rep","tech_rep"), sep=".", remove=TRUE) %>% 
  mutate(., Sample_name = case_when(Sample_name == "MiFish.Kangaroo.Run.3.NA.1" ~ "MiFish.Kangaroo.Run3",
                             TRUE ~ Sample_name))

MiFish4 <- read.csv(MiFish.asvs[9])
MiFish4$Sample_name <- as.character(MiFish4$Sample_name)
MiFish4 <- MiFish4 %>% select(-Locus) %>% left_join(all.metadata, by = "Sample_name") %>% select(Sample_ID,Locus,Hash,nReads) %>% rename("Sample_name" = "Sample_ID") %>% 
  separate(Sample_name, into=c("marker","date","creek","site","bio_rep","tech_rep"), remove = TRUE) %>%  # separate sample names 
  mutate(tech_rep = replace_na(tech_rep, 1)) %>% 
    mutate(., tech_rep = case_when(tech_rep == "TR2" ~ "2",
                                 tech_rep == "TR3" ~ "3",
                             TRUE ~ tech_rep)) %>% 
  unite("Sample_name", c("marker","date","creek","site","bio_rep","tech_rep"), sep=".", remove=TRUE) %>% 
  mutate(., Sample_name = case_when(Sample_name == "MiFish.Kangaroo.Run.4.NA.1" ~ "MiFish.Kangaroo.Run4",
                             TRUE ~ Sample_name))

MiFish5 <- read.csv(MiFish.asvs[10]) 
MiFish5 <- MiFish5 %>% 
  separate(Sample_name, into=c("marker","date","creek","site","bio_rep","tech_rep"), remove = TRUE) %>%  # separate sample names 
  mutate(tech_rep = replace_na(tech_rep, 1)) %>% 
    mutate(., tech_rep = case_when(tech_rep == "TR2" ~ "2",
                                 tech_rep == "TR3" ~ "3",
                             TRUE ~ tech_rep)) %>% 
  unite("Sample_name", c("marker","date","creek","site","bio_rep","tech_rep"), sep=".", remove=TRUE) %>% 
  mutate(., Sample_name = case_when(Sample_name == "MiFish.Kangaroo.Run.5.NA.1" ~ "MiFish.Kangaroo.Run5",
                             TRUE ~ Sample_name))

MiFish6 <- read.csv(MiFish.asvs[11])
MiFish6$Sample_name <- as.character(MiFish6$Sample_name)
MiFish6 <- MiFish6 %>% select(-Locus) %>% left_join(all.metadata, by = "Sample_name") %>% select(Sample_ID,Locus,Hash,nReads) %>% rename("Sample_name" = "Sample_ID") %>% 
  separate(Sample_name, into=c("marker","date","creek","site","bio_rep","tech_rep"), remove = TRUE) %>%  # separate sample names 
  mutate(tech_rep = replace_na(tech_rep, 1)) %>% 
    mutate(., tech_rep = case_when(tech_rep == "TR2" ~ "2",
                                 tech_rep == "TR3" ~ "3",
                             TRUE ~ tech_rep)) %>% 
  unite("Sample_name", c("marker","date","creek","site","bio_rep","tech_rep"), sep=".", remove=TRUE) %>% 
  mutate(., Sample_name = case_when(Sample_name == "MiFish.Kangaroo.Run.6.NA.1" ~ "MiFish.Kangaroo.Run6",
                             TRUE ~ Sample_name))

MiFish7 <- read.csv(MiFish.asvs[12])
MiFish7 <- MiFish7 %>% 
  separate(Sample_name, into=c("marker","date","creek","site","bio_rep","tech_rep"), remove = TRUE) %>%  # separate sample names 
  mutate(tech_rep = replace_na(tech_rep, 1)) %>% 
    mutate(., tech_rep = case_when(tech_rep == "TR2" ~ "2",
                                 tech_rep == "TR3" ~ "3",
                             TRUE ~ tech_rep)) %>% 
  unite("Sample_name", c("marker","date","creek","site","bio_rep","tech_rep"), sep=".", remove=TRUE) %>% 
  mutate(., Sample_name = case_when(Sample_name == "MiFish.Kangaroo.Run.7.NA.1" ~ "MiFish.Kangaroo.Run7",
                             TRUE ~ Sample_name))

MiFish8 <- read.csv(MiFish.asvs[13])
MiFish8 <- MiFish8 %>% 
  separate(Sample_name, into=c("marker","date","creek","site","bio_rep","tech_rep"), remove = TRUE) %>%  # separate sample names 
  mutate(tech_rep = replace_na(tech_rep, 1)) %>% 
    mutate(., tech_rep = case_when(tech_rep == "TR2" ~ "2",
                                 tech_rep == "TR3" ~ "3",
                             TRUE ~ tech_rep)) %>% 
  unite("Sample_name", c("marker","date","creek","site","bio_rep","tech_rep"), sep=".", remove=TRUE) %>% 
  mutate(., Sample_name = case_when(Sample_name == "MiFish.Kangaroo.NA.NA.NA.1" ~ "MiFish.Kangaroo.Run8",
                             TRUE ~ Sample_name))

MiFish9 <- read.csv(MiFish.asvs[14])
MiFish9 <- MiFish9 %>% 
  filter(!str_detect(Sample_name, "NEB")) %>% 
  filter(!str_detect(Sample_name, "ME")) %>% 
  separate(Sample_name, into=c("marker","date","creek","site","bio_rep","tech_rep"), remove = TRUE) %>%  # separate sample names 
  mutate(tech_rep = replace_na(tech_rep, 1)) %>% 
    mutate(., tech_rep = case_when(tech_rep == "TR2" ~ "2",
                                 tech_rep == "TR3" ~ "3",
                             TRUE ~ tech_rep)) %>% 
  unite("Sample_name", c("marker","date","creek","site","bio_rep","tech_rep"), sep=".", remove=TRUE) %>% 
  mutate(., Sample_name = case_when(Sample_name == "MiFish.Kangaroo.NA.NA.NA.1" ~ "MiFish.Kangaroo.Run9",
                             TRUE ~ Sample_name))

MiFish10 <- read.csv(MiFish.asvs[3])
MiFish10 <- MiFish10 %>% 
  separate(Sample_name, into=c("marker","date","creek","site","bio_rep","tech_rep"), remove = TRUE) %>%  # separate sample names 
  mutate(tech_rep = replace_na(tech_rep, 1)) %>% 
    mutate(., tech_rep = case_when(tech_rep == "TR2" ~ "2",
                                 tech_rep == "TR3" ~ "3",
                             TRUE ~ tech_rep)) %>% 
  unite("Sample_name", c("marker","date","creek","site","bio_rep","tech_rep"), sep=".", remove=TRUE) %>%
  mutate(., Sample_name = case_when(Sample_name == "MiFish.0222.Kangaroo.NA.NA.1" ~ "MiFish.Kangaroo.Run10",
                             TRUE ~ Sample_name))

MiFish11 <- read.csv(MiFish.asvs[4])
MiFish11 <- MiFish11 %>% 
  separate(Sample_name, into=c("marker","date","creek","site","bio_rep","tech_rep"), remove = TRUE) %>%  # separate sample names 
  mutate(tech_rep = replace_na(tech_rep, 1)) %>% 
    mutate(., tech_rep = case_when(tech_rep == "TR2" ~ "2",
                                 tech_rep == "TR3" ~ "3",
                             TRUE ~ tech_rep)) %>% 
  unite("Sample_name", c("marker","date","creek","site","bio_rep","tech_rep"), sep=".", remove=TRUE) %>% 
  mutate(., Sample_name = case_when(Sample_name == "MiFish.0222.Kangaroo.NA.NA.1" ~ "MiFish.Kangaroo.Run11",
                             TRUE ~ Sample_name))

MiFish12 <- read.csv(MiFish.asvs[5])
MiFish12 <- MiFish12 %>% 
  separate(Sample_name, into=c("marker","date","creek","site","bio_rep","tech_rep"), remove = TRUE) %>%  # separate sample names 
  mutate(tech_rep = replace_na(tech_rep, 1)) %>% 
  mutate(., tech_rep = case_when(tech_rep == "TR2" ~ "2",
                                 tech_rep == "TR3" ~ "3",
                             TRUE ~ tech_rep)) %>% 
  unite("Sample_name", c("marker","date","creek","site","bio_rep","tech_rep"), sep=".", remove=TRUE) %>% 
  mutate(., Sample_name = case_when(Sample_name == "MiFish.Kangaroo.NA.NA.NA.1" ~ "MiFish.Kangaroo.Run12",
                             TRUE ~ Sample_name))



MiFish.ASV.table <- bind_rows(MiFish2, MiFish3, MiFish4, MiFish5, MiFish6, MiFish7, MiFish8, MiFish9, MiFish10, MiFish11, MiFish12)

MiFish.hash.key <- bind_rows(map(MiFish.hashes, read_csv))
MiFish.hash.key <- MiFish.hash.key %>% 
  distinct(Hash, .keep_all = T) 

MiFish.ASV.table <- MiFish.ASV.table %>% 
   mutate(., Sample_name = case_when(Sample_name == "MiFish.0321.1Prt.Dn.1TR2.1" ~ "MiFish.0321.1Prt.Dn.1.2",
                                     Sample_name == "MiFish.0321.1Prt.Dn.1TR3.1" ~ "MiFish.0321.1Prt.Dn.1.3",
                             Sample_name == "MiFish.0421.3Chk.Dn.3TR2.1" ~ "MiFish.0421.3Chk.Dn.3.2",
                             Sample_name == "MiFish.0421.3Chk.Dn.3TR3.1" ~ "MiFish.0421.3Chk.Dn.3.3",
                             Sample_name == "MiFish.0821.2Brn.Dn.2TR2.1" ~ "MiFish.0821.2Brn.Dn.2.2",
                             Sample_name == "MiFish.0821.2Brn.Dn.2TR3.1" ~ "MiFish.0821.2Brn.Dn.2.3",
                             TRUE ~ Sample_name)) 

```

## Merge MiMammal runs

```{r pull and merge ASV tables and hashes MiMammal} 

# mimammal runs 1 and 4 have sample as NGNX rather than sample name so have to fix
MiMammal.ASV.table <- bind_rows(map(MiMammal.asvs[2:3], read_csv))
MiMammal1 <- read.csv(MiMammal.asvs[1])
MiMammal1$Sample_name <- as.character(MiMammal1$Sample_name)
MiMammal1 <- MiMammal1 %>% select(-Locus) %>% left_join(all.metadata, by = "Sample_name") %>% select(Sample_ID,Locus,Hash,nReads) %>% rename("Sample_name" = "Sample_ID")
MiMammal4 <- read.csv(MiMammal.asvs[4])
MiMammal4$Sample_name <- as.character(MiMammal4$Sample_name)
MiMammal4 <- MiMammal4 %>% select(-Locus) %>% left_join(all.metadata, by = "Sample_name") %>% select(Sample_ID,Locus,Hash,nReads) %>% rename("Sample_name" = "Sample_ID")
MiMammal5 <- read.csv(MiMammal.asvs[5])
MiMammal6 <- read.csv(MiMammal.asvs[6])
MiMammal6$Sample_name <- as.character(MiMammal6$Sample_name)
MiMammal6 <- MiMammal6 %>% select(-Locus) %>% left_join(all.metadata, by = "Sample_name") %>% select(Sample_ID,Locus,Hash,nReads) %>% rename("Sample_name" = "Sample_ID")
MiMammal7 <- read.csv(MiMammal.asvs[7])
MiMammal7 <- MiMammal7 %>% 
  separate(Sample_name, into=c("marker","date","creek","site","bio_rep"), remove = TRUE) %>%  # separate sample names 
  unite("Sample_name", c("marker","date","creek","site","bio_rep"), sep=".", remove=TRUE)


MiMammal.ASV.table <- bind_rows(MiMammal.ASV.table, MiMammal1, MiMammal4, MiMammal5, MiMammal6, MiMammal7)

MiMammal.hash.key <- bind_rows(map(MiMammal.hashes, read_csv))
MiMammal.hash.key <- MiMammal.hash.key %>% 
  distinct(Hash, .keep_all = T) 

```

# Write all the merged files for all markers

```{r write combined files}

# Write the Hash keys 
COI.hash.key %>% 
  write_csv(paste0(here("Output", "dada2_output"), "/", format(Sys.Date(), "%Y%m%d") , ".combined.COI.hash.key.csv"))
MiFish.hash.key %>% 
  write_csv(paste0(here("Output", "dada2_output"), "/", format(Sys.Date(), "%Y%m%d") , ".combined.MiFish.hash.key.csv"))
MiMammal.hash.key %>% 
  write_csv(paste0(here("Output", "dada2_output"), "/", format(Sys.Date(), "%Y%m%d") , ".combined.MiMammal.hash.key.csv"))

# Write the merged ASV tables
COI.ASV.table %>% 
  write_csv(paste0(here("Output", "dada2_output"), "/", format(Sys.Date(), "%Y%m%d") , ".combined.COI.ASV.table.csv"))
MiFish.ASV.table %>% 
  write_csv(paste0(here("Output", "dada2_output"), "/", format(Sys.Date(), "%Y%m%d") , ".combined.MiFish.ASV.table.csv"))
MiMammal.ASV.table %>% 
  write_csv(paste0(here("Output", "dada2_output"), "/", format(Sys.Date(), "%Y%m%d") , ".combined.MiMammal.ASV.table.csv"))

```
